Add comments and a fuzz test for `flipLiquidityDelta` regarding `type(int128).min`
This commit refactors the `flipLiquidityDelta` function in `LiquidityMath` library that can now handle `type(int128).min` and returns a valid `int256`. It also modifies the use of `flipLiquidityDelta` in `Pool` library along with improved explanations with inline comments. Furthermore, it updates the tests in `LiquidityMath.t.sol` by refining the fuzzing, modifying import statements and introducing a new test case for `flipLiquidityDelta` handling `type(int128).min`.

+++ src/libraries/LiquidityMath.sol
@@ -22,11 +22,11 @@ library LiquidityMath {
 
     /// @notice Flips the sign of a liquidity delta if a condition is true
     /// @dev Equivalent to `res = flip ? -liquidityDelta : liquidityDelta;`
-    /// @dev `liquidityDelta` should never be `type(int128).min`
+    /// @dev If flipping `type(int128).min`, `1 << 127` is returned which is a valid `int256`
     /// @param liquidityDelta The liquidity delta to potentially flip
     /// @param flip Whether to flip the sign of the liquidity delta
     /// @return res The potentially flipped liquidity delta
-    function flipLiquidityDelta(int128 liquidityDelta, bool flip) internal pure returns (int128 res) {
+    function flipLiquidityDelta(int128 liquidityDelta, bool flip) internal pure returns (int256 res) {
         assembly {
             // if flip = true, res = -liquidityDelta = ~liquidityDelta + 1 = (-1) ^ liquidityDelta + 1
             // if flip = false, res = liquidityDelta = 0 ^ liquidityDelta + 0
